Reduce set configuration validation error logs
