+++ src/main/java/proai/driver/daos/json/SetSpecDaoJson.java
@@ -23,7 +23,44 @@ public class SetSpecDaoJson {
 
     public SetSpecDaoJson(InputStream setSpecsConfiguration) throws IOException {
         ObjectMapper om = new ObjectMapper();
-        sets = om.readValue(setSpecsConfiguration, om.getTypeFactory().constructCollectionType(List.class, Set.class));
+        sets = om.readValue(
+                setSpecsConfiguration,
+                om.getTypeFactory().constructCollectionType(List.class, Set.class));
+        if (!validate(sets)) {
+            logger.warn("There are problems with the setSpec configuration. See log for details.");
+        }
+    }
+
+    private boolean validate(List<Set> sets) {
+        boolean result = true;
+        for (Set set : sets) {
+            String setSpec = set.getSetSpec();
+
+            if (assertNotNullNotEmpty(setSpec, "Found empty setSpec")) {
+                result &= false;
+            } else {
+
+                String setName = set.getSetName();
+                if (assertNotNullNotEmpty(setName, String.format("Found empty setName for %s", setSpec))) {
+                    result &= false;
+                }
+
+                String setPredicate = set.getPredicate();
+                if (assertNotNullNotEmpty(setPredicate, String.format("Found empty set predicate for '%s'.", setSpec))) {
+                    result &= false;
+                }
+
+            }
+        }
+        return result;
+    }
+
+    private boolean assertNotNullNotEmpty(String setName, String s) {
+        if (setName == null || setName.isEmpty()) {
+            logger.warn(s);
+            return true;
+        }
+        return false;
     }
 
     public List<Set> getSetObjects() {
